mm/mm_init.c: avoid deferred init on MEMBLOCK_SCRATCH pages
The MEMBLOCK_SCRATCH pages need to be marked as CMA before the pages are
handed off to the buddy allocator. To do that, the pages must be
initialized. If that is not done, a VM_BUG_ON() is hit when
CONFIG_DEBUG_VM is enabled. Without CONFIG_DEBUG_VM, the pages will
silently be handed over to the buddy allocator by page_alloc_init_late()
since it thinks those pages are not yet initialized.

Fix that by avoiding deferred allocation on MEMBLOCK_SCRATCH pages.

Signed-off-by: Pratyush Yadav <ptyadav@amazon.de>