Introduce a super crude slab allocator
Generally, the kernel slab cache allocator is used to allocate various
kind of objects. But if we want to persist those with KHO, we need to
find a way to serialize/deserialize the state of the slab cache. That is
a very complex task and would take a long time.

So for now, write a super simple slab allocator that allocates memory in
2 MiB blocks and hands out slots of requested size. It is fully
serialized by a mutex and does not even know how to free its backing
memory. Still, it is good enough to get going with a prototype. It
doesn't deserve much attention since it will not be used for "real"
patches anyway.

NOTE: As with every patch is this series, the code here is very hacky
and quickly put together to get a proof-of-concept. So the code is full
of bugs and problems, and will be significantly improved before the
first round of patches on the list.

Signed-off-by: Pratyush Yadav <ptyadav@amazon.de>