Add TMPFS_KHO config and option
Add a Kconfig to enable/disable KHO support for TMPFS. In later parts of
the code, it isn't actually being used to it doesn't do much for now.
Still, later KHO code should be wrapped around a check for this.

Also add a mount option called "kho" to TMPFS. It should be specified by
a string that identifies the mount at later boots. Just store the
information for now, later patches will use it.

NOTE: As with every patch is this series, the code here is very hacky
and quickly put together to get a proof-of-concept. So the code is full
of bugs and problems, and will be significantly improved before the
first round of patches on the list.

Signed-off-by: Pratyush Yadav <ptyadav@amazon.de>