Introduce KHO inode
When using KHO, a stable inode format is needed that can be shared
between different kernel versions. To achieve that, a simpler, stable
version of the TMPFS/SHMEM inode is created here. We don't care about
endianness here since the inode will always be used on the same machine.

The first parts of the inode contain things common with many other Linux
file systems: mode, uid, gid, atime, mtime, ctime, size, etc. These
should largely be used as-is and should not be too interesting.

The more interesting bits are the ones I haven't figured out yet: how to
we track data blocks, both for regular files and directory entries or
symlinks. We can take ext2's simple table-based scheme, or do something
more complex with B-trees or maybe even something resembling xarray. I
have not thought deeply about this yet.

For now, the inode just contains a flat array that tracks where each
block is located for regular files. The array has 2036 entries which
makes the inode an even 16 KiB. This means each file can only have a
maximum of around 7.9 MiB of data. Not too useful in real life, but
serves as a proof of concept.

The big array also serves to hold directory entries. Each directory
entry is a 264 byte structure with 8 bytes for inode number, one byte
for file type, and 255 bytes (including NUL-terminator) for name.
Currently, directory entries can only be stored in-inode and can not be
stored in external blocks, so the maximum number of directory entries
adds up to 61.

The inode numbers currently are exactly equal to the virtual address of
the KHO inode. This of course has multiple problems like leaking kernel
memory address, not working with KASLR, etc. but works to easily and
quickly number the inodes without having to some up with something
smarter.

NOTE: As with every patch is this series, the code here is very hacky
and quickly put together to get a proof-of-concept. So the code is full
of bugs and problems, and will be significantly improved before the
first round of patches on the list.

Signed-off-by: Pratyush Yadav <ptyadav@amazon.de>