Introduce {freeze,thaw}_super_locked()
I plan to use them when serializing/deserializing the superblock to
freeze and thaw the superblocks, but as of now haven't really used them
in my code since the locking is tiny bit complex and I haven't managed
to figure it out yet. Still, leaving this around in case I want to use
it later.

NOTE: As with every patch is this series, the code here is very hacky
and quickly put together to get a proof-of-concept. So the code is full
of bugs and problems, and will be significantly improved before the
first round of patches on the list.

Signed-off-by: Pratyush Yadav <ptyadav@amazon.de>