Add ability to iterate superblocks in more ways
First, add ability to specify lock type: exclusive or shared. Second,
add ability to return errors from the callbacks and bail if any call
does return non-zero.

These will be used when serializing the superblocks before KHO to
iterate all TMPFS/SHMEM superblocks and find the ones that are KHO
mounts.

Do some refactoring of iterate_supers_type() to make implementing the
shared/exclusive locking easier by introducing a helper.

NOTE: As with every patch is this series, the code here is very hacky
and quickly put together to get a proof-of-concept. So the code is full
of bugs and problems, and will be significantly improved before the
first round of patches on the list.

Signed-off-by: Pratyush Yadav <ptyadav@amazon.de>