Introduce bitmap_kv[z]alloc() and bitmap_kvfree()
When allocating large bitmaps, kmalloc() might not be suitable since it
is recommended only for allocations smaller then page size. Allow
allocating bitmaps using kvalloc() to allow larger allocations to
succeed.

KHO code in SHMEM will allocate bitmaps with one bit for each PFN. These
APIs will be used by it since on machines with more than 16 MiB of RAM,
the bitmap will span more than a page.

NOTE: As with every patch is this series, the code here is very hacky
and quickly put together to get a proof-of-concept. So the code is full
of bugs and problems, and will be significantly improved before the
first round of patches on the list.

Signed-off-by: Pratyush Yadav <ptyadav@amazon.de>