Introduce the bulk of TMPFS-KHO ser-des and inode management logic
This patch contains a big chunk of changes that add the logic for
creating and managing a KHO inode, a KHO superblock, along with logic to
serialize and deserialize the inodes and superblocks via KHO. This is
all done in a bit patch mainly because I don't want to spend too much
time right now to break it down into smaller chunks, since these all are
somewhat closely related with each other.

On the inode side, this patch adds a reference to a kho_inode to
shmem_inode_info. The KHO inode itself is allocated from a simple_slab
allocator, which is then serialized-deserialized via KHO, carrying the
inodes across kexec.

shmem_get_inode() is modified to also allocate and initialize a KHO
inode for KHO mounts. The KHO inode is initialized based on the in-core
inode and on later can be used to re-create the in-core inode. This
re-creation is done via shmem_get_kho_inode(). Some refactoring is done
and some helpers added to make this easier.

On the superblock side, shmem_fill_super() is modified to initialize the
bitmap of all pages that should be preserved via KHO. This is of course
inefficient, but a better scheme can be derived later. It also sets up
other KHO related metadata on a first-mount, or deserialized old
metadata on a post-kexec-mount. Some flags are added to shmem_sb_info to
track some states like whether this is a KHO-enabled mount, and whether
this mount came from a KHO deserialization.

On serialize side, shmem_kho_serialize() takes care of serializing the
superblocks. It does so by iterating over all SHMEM superblocks and
checking which ones are KHO mounts. The function is also supposed to
freeze the superblocks before serializing them, but I have not
implemented that yet since the locking with superblocks is a tiny bit
tricky and I didn't spend the time to figure it out.

On serialize, we save the KHO inode cache, all pages used by files, and
metadata for all superblocks. Each superblock is serialized via
shmem_serialize_super(). It saves metadata like uid, gid, etc. along
with flags and root inode number.

On deserialize, we restore common things like inode cache and file
pages, and leave the rest for later when the superblock is mounted. Upon
superblock mount, we restore metadata like uid, gid, etc., along with
the root inode. The other inodes will be restored on-demand in later
patches.

An example of a serialized devicetree:

/dts-v1/;

/ {
        compatible = "kho-v1";

        tmpfs {
                compatible = "tmpfs-kho-v1";
                mem = <0xd0ce12 0x1000000 0x100000 0x00>;

                inodes {
                        compatible = "simple-slab-v1";
                        mem = <0x16 0x1000000 0x2000 0x00>;
                        slabsize = <0x00 0x4000>;
                        first-free = <0xffff8881 0x16040000>;
                };

                supers {

                        test {
                                compatible = "tmpfs-kho-super-v1";
                                uid = <0x00>;
                                gid = <0x00>;
                                mode = <0x3ff>;
                                max-blocks = <0x00 0x1f3d51>;
                                max-inodes = <0x00 0x1f3d51>;
                                free-ispace = <0x00 0x7cf54400>;
                                flags = <0x00 0x70810000>;
                                root = <0xffff8881 0x16000000>;
                        };
                };
        };
};

NOTE: As with every patch is this series, the code here is very hacky
and quickly put together to get a proof-of-concept. So the code is full
of bugs and problems, and will be significantly improved before the
first round of patches on the list.

Signed-off-by: Pratyush Yadav <ptyadav@amazon.de>